pyproject.toml
[project]
name = "godzkilla"
version = "0.2.0"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9",
    "hayeah-core",
]

[project.scripts]
godzkilla = "godzkilla.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/godzkilla"]

[tool.uv.sources]
hayeah-core = { path = "../dotfiles/hayeah", editable = true }

[dependency-groups]
dev = [
    "pytest",
]

[tool.pytest.ini_options]
testpaths = ["src"]

[tool.pyright]
pythonVersion = "3.11"
include = ["src"]
typeCheckingMode = "basic"
